Various parsing fixes.

+++ d11-boot-parser/src/main/java/org/d11/boot/parser/match/whoscored/v1/model/IncidentEvent.java
@@ -49,7 +49,7 @@ public class IncidentEvent {
     /**
      * Id of the player involved in the event.
      */
-    private long playerId;
+    private int playerId;
     /**
      * Related event id. For example a substitution on <-> substitution off.
      */
